// src/app/shared/models/Interfaces/objeto.ts
import { Comentario } from "./comentario";
import { Anexo } from "./anexo";
import { Coordenadas } from "./coordenadas";
import { HistorialNombresHistoricos } from "./historialNombresHistoricos";
import { HistorialNombresLocales } from "./historialNombresLocales";
import { Poblacion } from "./poblacion";

export interface Objeto {
  id:                      number;
  claseObjetoId:           number;
  subClaseObjetoId:        number;
  estadoId:                number;
  regionId:                number;
  provinciaId:             number;
  municipioId:             number;
  distritoMunicipalId:     number;
  codigo:                  string;
  codigoOne:               string;
  nombreGeografico:        string;
  nombreLocal:             string;
  nombreHistorico:string;
  historia:                string;
  area:                    string;
  altitud:                 string;
  longitud:                string;
  coordenadaUtmX:string;
  coordenadaUtmY:string;
  coordenadaGeoX:string;
  coordenadaGeoY:string;
  divisionPolitica:        string;
  ficha:                   string;
  linkFicha:               string;
  numeroSerie:             string;
  baseLegal:               string;
  estatus:                 boolean;
  borrado:                 boolean;
  creadoPor:               number;
  modificadoPor:           number;
  nombreCreadoPor:         string;
  nombreModificadoPor:     string;
  fechaRegistro:           Date;
  fechaModificacion:       Date;
  nombreClaseObjeto:       string;
  nombreSubClaseObjeto:    string;
  nombreEstado:            string;
  permisoPorEstado:        string;
  nombreRegion:            string;
  nombreProvincia:         string;
  nombreMunicipio:         string;
  nombreDistritoMunicipal: string;
  seccionId?: number;
  nombreSeccion:string;
  parajeId?: number;
  nombreParaje:string;
  barrioId?: number;
  nombreBarrio:string;
  subBarrioId?: number;
  nombreSubBarrio:string;
  imagenesCount:           number;
  documentosCount:         number;
  usaurio:                 null;
  anexos:                  Anexo[];
  anexosCount:             number;
  coordenadas:Coordenadas[];
  coordenadasCount:number;
  poblaciones:Poblacion[];
  poblacionesCount:number;
  historialNombresLocales:HistorialNombresLocales[];
  historialNombresLocalesCount:number;
  historialNombresHistoricos:HistorialNombresHistoricos[];
  historialNombresHistoricosCount:number;
  tieneMunicipio: boolean;
  tieneDistritoMunicipal: boolean;
  tieneSeccion: boolean;
  tieneParaje: boolean;
  tieneBarrio: boolean;
  tieneSubBarrio: boolean
}